backend/requirements.txt
Flask
flask-cors
matplotlib
memory_profiler
numpy
Pillow
tqdm